requirements.txt
Django==3.2.9
django-cors-headers==3.4.0
django-environ==0.4.5
django-storages==1.9.1
djangorestframework==3.11.0
dropbox==10.2.0
gunicorn==20.0.4
Pillow==7.1.2
psycopg2==2.8.5
requests==2.24.0
whitenoise==5.1.0
